+++ src/writer/writer-ts-definitions-core.ts
@@ -765,17 +765,17 @@ function genModuleExports(def: Pick<ComponentDocApi, "moduleExports">) {
 
 export function writeTsDefinition(component: ComponentDocApi) {
   const {
-    moduleName,
-    typedefs,
+    moduleName = "",
+    typedefs = [],
     generics,
-    props,
-    moduleExports,
-    slots,
-    events,
+    props = [],
+    moduleExports = [],
+    slots = [],
+    events = [],
     rest_props,
     extends: _extends,
     componentComment,
-    contexts,
+    contexts = [],
   } = component;
   const { props_name, prop_def } = genPropDef({
     moduleName,
@@ -800,9 +800,12 @@ export function writeTsDefinition(component: ComponentDocApi) {
   const needsHTMLAttributes =
     rest_props?.type === "Element" && rest_props.name === "svelte:element" && !rest_props.thisValue;
 
+  const needsSnippet = props?.some((p) => p.type?.includes("Snippet")) ?? false;
+
   return `
   import { SvelteComponentTyped } from "svelte";${
-    needsSvelteHTMLElements ? `import type { SvelteHTMLElements } from "svelte/elements";\n` : ""
+    needsSnippet ? `import type { Snippet } from "svelte";\n` : ""
+  }${needsSvelteHTMLElements ? `import type { SvelteHTMLElements } from "svelte/elements";\n` : ""
   }${needsHTMLAttributes ? `import type { HTMLAttributes } from "svelte/elements";\n` : ""}
   ${genImports({ extends: _extends })}
   ${genModuleExports({ moduleExports })}