+++ tests/e2e/carbon/rollup.config.js
@@ -2,7 +2,7 @@ import pkg from "./package.json";
 import svelte from "rollup-plugin-svelte";
 import resolve from "@rollup/plugin-node-resolve";
 import commonjs from "@rollup/plugin-commonjs";
-import sveld from "sveld";
+import veld from "veld";
 
 const production = !process.env.ROLLUP_WATCH;
 
@@ -14,7 +14,7 @@ export default {
     resolve(),
     commonjs(),
     production &&
-      sveld({
+      veld({
         glob: true,
         typesOptions: {
           preamble: `// TypeScript definitions for ${pkg.name}@${pkg.version}\n\n`,