+++ tests/e2e/multi-export-typed-ts-only/rollup.config.js
@@ -1,7 +1,7 @@
 import pkg from "./package.json";
 import svelte from "rollup-plugin-svelte";
 import resolve from "@rollup/plugin-node-resolve";
-import sveld from "sveld";
+import veld from "veld";
 
 const production = !process.env.ROLLUP_WATCH;
 
@@ -11,6 +11,6 @@ export default {
   plugins: [
     svelte(),
     resolve(),
-    production && sveld(),
+    production && veld(),
   ],
 };