+++ tests/e2e/path-aliases/rollup.config.js
@@ -1,7 +1,7 @@
 import pkg from "./package.json";
 import svelte from "rollup-plugin-svelte";
 import resolve from "@rollup/plugin-node-resolve";
-import sveld from "sveld";
+import veld from "veld";
 
 const production = !process.env.ROLLUP_WATCH;
 
@@ -12,7 +12,7 @@ export default {
     svelte(),
     resolve(),
     production &&
-      sveld({
+      veld({
         markdown: true,
         markdownOptions: {
           onAppend: (type, document, components) => {